rename: ContentsBox 에서 사용하는 Component 들 폴더 구조 변경
